+++ doc/bufferline.txt
@@ -12,6 +12,7 @@ Introduction...........................: |bufferline-introduction|
 Usage..................................: |bufferline-usage|
 Settings...............................: |bufferline-settings|
 Styling................................: |bufferline-styling|
+Tabpages...............................: |bufferline-tabpages|
 Numbers................................: |bufferline-numbers|
 LSP Diagnostics........................: |bufferline-diagnostics|
 Groups.................................: |bufferline-groups|
@@ -29,7 +30,7 @@ Issues.................................: |bufferline-issues|
 ==============================================================================
 INTRODUCTION				        *bufferline-introduction*
 
-A _snazzy_ 💅 buffer line (with minimal tab integration) for Neovim built
+A _snazzy_ 💅 buffer line (tab integration) for Neovim built
 using `lua`.
 
 This plugin shamelessly attempts to emulate the aesthetics of GUI text
@@ -53,6 +54,7 @@ The available settings are: >
 
     require"bufferline".setup{
         options = {
+            mode = "buffers", -- can also be set to "tabs" to see tabpages
             numbers = "none" | "ordinal" | "buffer_id" | "both" | function({ ordinal, id, lower, raise }): string,
             --- @deprecated, please specify numbers as a function to customize the styling
             number_style = "superscript" | "" | { "none", "subscript" }, -- buffer_id at index 1, ordinal at index 2
@@ -115,6 +117,18 @@ You can change the appearance of the bufferline separators by setting the
 * finally you can pass in a custom list containing 2 characters which will be
   used as the separators e.g. `{"|", "|"}`, the first is the left and the
   second is the right separator
+
+
+==============================================================================
+TABPAGES                                                          *bufferline-tabs*
+
+This plugin can also be set to show only tabpages. This can be done by setting the `mode` option to `tabs`. This will change the bufferline to a tabline
+it has a lot of the same features/styling but not all.
+A few things to note are
+
+* Diagnostics only show if the buffer with issues is the current window selected in that tab page
+* Sorting doesn't work yet as that needs to be thought through.
+* Grouping doesn't work yet as that also needs to be thought through.
 
 
 ==============================================================================
