+++ lua/bufferline/diagnostics.lua
@@ -49,7 +49,6 @@ local mt = {
   end,
 }
 
-local is_nightly = utils.is_truthy(fn.has("nvim-0.7"))
 local is_valid_version = utils.is_truthy(fn.has("nvim-0.5"))
 
 local function is_disabled(diagnostics)
@@ -151,22 +150,22 @@ end
 
 ---@param context RenderContext
 function M.component(context)
-  local opts = context.preferences.options
+  local opts = require("bufferline.config").get("options")
   if is_disabled(opts.diagnostics) then
     return context
   end
 
   local user_indicator = opts.diagnostics_indicator
   local highlights = context.current_highlights
-  local buf = context.tab:as_buffer()
-  local diagnostics = buf.diagnostics
-  if diagnostics.count < 1 then
+  local element = context.tab
+  local diagnostics = element.diagnostics
+  if not diagnostics or not diagnostics.count or diagnostics.count < 1 then
     return context
   end
 
   local indicator = " (" .. diagnostics.count .. ")"
   if user_indicator and type(user_indicator) == "function" then
-    local ctx = { buffer = buf }
+    local ctx = { buffer = element, tab = element }
     indicator = user_indicator(diagnostics.count, diagnostics.level, diagnostics.errors, ctx)
   end
