+++ lua/bufferline/pick.lua
@@ -10,20 +10,20 @@ function M.reset()
   M.current = {}
 end
 
----@param buf Buffer
+---@param element Tabpage|Buffer
 ---@return string?
-function M.get(buf)
-  local first_letter = buf.filename:sub(1, 1)
+function M.get(element)
+  local first_letter = element.name:sub(1, 1)
   -- should only match alphanumeric characters
   local invalid_char = first_letter:match("[^%w]")
 
   if not M.current[first_letter] and not invalid_char then
-    M.current[first_letter] = buf.id
+    M.current[first_letter] = element.id
     return first_letter
   end
   for letter in valid:gmatch(".") do
     if not M.current[letter] then
-      M.current[letter] = buf.id
+      M.current[letter] = element.id
       return letter
     end
   end
@@ -31,20 +31,20 @@ end
 
 ---@param ctx RenderContext
 function M.component(ctx)
-  local padding = require("bufferline.constants").padding
   local utils = require("bufferline.utils")
+  local padding = require("bufferline.constants").padding
+  local options = require("bufferline.config").get("options")
 
-  local buffer = ctx.tab:as_buffer()
+  local element = ctx.tab
   local length = ctx.length
   local component = ctx.component
-  local options = ctx.preferences.options
   local hl = ctx.current_highlights
-  local letter = buffer.letter
+  local letter = element.letter
 
-  if options.show_buffer_icons and buffer.icon then
-    local right = string.rep(padding, math.ceil((strwidth(buffer.icon) - 1) / 2))
-    local left = string.rep(padding, math.floor((strwidth(buffer.icon) - 1) / 2))
-    letter = left .. buffer.letter .. right
+  if options.show_buffer_icons and element.icon then
+    local right = string.rep(padding, math.ceil((strwidth(element.icon) - 1) / 2))
+    local left = string.rep(padding, math.floor((strwidth(element.icon) - 1) / 2))
+    letter = left .. element.letter .. right
   end
 
   component = utils.join(hl.pick, letter, padding, hl.background, component)
