+++ lua/bufferline/sorters.lua
@@ -1,3 +1,5 @@
+local config = require("bufferline.config")
+
 local M = {}
 ---------------------------------------------------------------------------//
 -- Sorters
@@ -17,7 +19,7 @@ end
 --- @param buf_a Buffer
 --- @param buf_b Buffer
 local function sort_by_extension(buf_a, buf_b)
-  return fnamemodify(buf_a.filename, ":e") < fnamemodify(buf_b.filename, ":e")
+  return fnamemodify(buf_a.name, ":e") < fnamemodify(buf_b.name, ":e")
 end
 
 --- @param buf_a Buffer
@@ -89,25 +91,28 @@ local function sort_by_tabs(buf_a, buf_b)
 end
 
 --- sorts a list of buffers in place
---- @param sort_by string|function
---- @param buffers Buffer[]
-function M.sort_buffers(sort_by, buffers)
+--- @param elements TabElement[]
+function M.sort(elements)
+  local sort_by = config.options.sort_by
+  local is_tabline = config:is_tabline()
+
   if sort_by == "extension" then
-    table.sort(buffers, sort_by_extension)
+    table.sort(elements, sort_by_extension)
   elseif sort_by == "directory" then
-    table.sort(buffers, sort_by_directory)
+    table.sort(elements, sort_by_directory)
   elseif sort_by == "relative_directory" then
-    table.sort(buffers, sort_by_relative_directory)
+    table.sort(elements, sort_by_relative_directory)
   elseif sort_by == "id" then
-    table.sort(buffers, sort_by_id)
+    table.sort(elements, sort_by_id)
   elseif sort_by == "tabs" then
-    table.sort(buffers, sort_by_tabs)
+    table.sort(elements, is_tabline and sort_by_id or sort_by_tabs)
   elseif type(sort_by) == "function" then
-    table.sort(buffers, sort_by)
+    table.sort(elements, sort_by)
   end
-  for index, buf in ipairs(buffers) do
+  for index, buf in ipairs(elements) do
     buf.ordinal = index
   end
+  return elements
 end
 
 return M
