+++ tests/sorters_spec.lua
@@ -1,9 +1,16 @@
 describe("Sorters - ", function()
   local sorters = require("bufferline.sorters")
+  local bufferline
+
+  before_each(function()
+    package.loaded["bufferline"] = nil
+    bufferline = require("bufferline")
+  end)
 
   it("should sort by ID correctly", function()
+    bufferline.setup({ options = { sort_by = "id" } })
     local bufs = { { id = 12 }, { id = 2 }, { id = 3 }, { id = 8 } }
-    sorters.sort_buffers("id", bufs)
+    sorters.sort(bufs)
     local ids = vim.tbl_map(function(buf)
       return buf.id
     end, bufs)
@@ -11,6 +18,7 @@ describe("Sorters - ", function()
   end)
 
   it("should sort by components correctly", function()
+    bufferline.setup({ options = { sort_by = "tabs" } })
     vim.cmd("e file1.txt")
     vim.cmd("tabnew file2.txt")
     vim.cmd("tabnew file3.txt")
@@ -19,7 +27,7 @@ describe("Sorters - ", function()
       return { id = id }
     end, vim.api.nvim_list_bufs())
 
-    sorters.sort_buffers("tabs", bufs)
+    sorters.sort(bufs)
 
     local buf_names = vim.tbl_map(function(buf)
       return vim.fn.fnamemodify(vim.api.nvim_buf_get_name(buf.id), ":p:t")
